+++ docs/app/pages/components/components-sections/dashboard/widgets/dashboard-widgets.component.html
@@ -11,7 +11,7 @@
         [status]="status"
         [actions]="actions">
         <ng-template #iconAccept>
-            <span class="custom-icon">✅ Accept</span>
+            <span class="custom-icon">✅</span>
         </ng-template>
     </ux-dashboard-actions-widget>
