+++ src/components/dashboard-widgets/interfaces/actions-widget.interface.ts
@@ -6,7 +6,7 @@ export interface ActionsWidgetConfig {
 }
 
 export interface ActionConfig {
-    label?: string;
+    label: string;
     icon?: string;
     iconTemplate?: TemplateRef<any>;
     action: Function;
