+++ src/components/dashboard-widgets/interfaces/predefined-widget.interface.ts
@@ -2,8 +2,8 @@ export interface PredefinedWidgetConfig {
     id: string;
     name: string;
     heading: string;
-    fixedMode: boolean;
+    fixedMode?: boolean;
     resizable?: boolean;
-    colSpan: number;
-    rowSpan: number;
+    colSpan?: number;
+    rowSpan?: number;
 }
