+++ src/components/dashboard-widgets/interfaces/table-widget.interface.ts
@@ -1,5 +1,5 @@
 export interface TableWidgetConfig {
-    header: ReadonlyArray<any>;
+    header?: ReadonlyArray<any>;
     data: ReadonlyArray<ReadonlyArray<any>>;
     resizable?: boolean;
 }
