+++ src/components/dashboard-widgets/interfaces/text-widget.interface.ts
@@ -1,0 +1,5 @@
+export interface TextWidgetConfig {
+    text: string;
+    editable?: boolean;
+    resizable?: boolean;
+}
