--- src/components/dashboard-widgets/interfaces/text-widget.ts
@@ -1,13 +1,0 @@
-import { ElementRef } from '@angular/core';
-import { SidePanelComponent } from '../../side-panel';
-
-export interface TextWidgetConfig {
-    text: string;
-    editable?: boolean;
-    resizable?: boolean;
-    sidePanel: SidePanelComponent;
-    textArea: ElementRef<HTMLTextAreaElement>;
-    open: () => void;
-    cancel: () => void;
-    save: () => void;
-}
